==> 2 install packages.R <==
setwd("/Users/...")
update.packages(ask = FALSE, dependencies = c ( 'Suggests' ))

library(gplots)
library(ROCR)
library(MatrixModels)
library(foreign)
library(haven)
library(Biograph)
library(survival)
library(sp)
library(rworldmap)
library(ggplot2)
library(ggmap)
library(data.table)
library(plyr)
library(dplyr)
library(Matrix)
install.packages("INLA", repos="https://inla.r-inla-download.org/R/testing", dep=TRUE)
install.packages("INLA", repos="https://inla.r-inla-download.org/R/stable", dep=TRUE)
install.packages("INLA", repos=c(getOption("repos"), INLA="https://inla.r-inla-download.org/R/stable"), dep=TRUE)
update.packages("INLA", dep=TRUE)
library(parallel)
library(INLA)
library(bdsmatrix)
library(coxme)
install.packages("devtools")
library(usethis)
library(devtools)
install_github('timcdlucas/INLAutils')
library(INLAutils)
library(maps)
library(mapdata)
library(graphics)
library(spacetime)
library(plotGoogleMaps)
library(stats)
library(reshape2)
library(maptools)
library(rgdal)
library(DBI)
library(RPostgreSQL)
library(sf)
library(maptools)
library(spData)
library(spdep)
library(scales)
library(gdata)
library(questionr)
inla.setOption("enable.inla.argument.weights", TRUE) 
install_github("julianfaraway/brinla")
library(brinla)
library(rgdal)
library(maptools)
library(gpclib)
if (!require(gpclib)) install.packages("gpclib", type="source")
gpclibPermit()
library(caTools)
library(gplots)
library(plotrix)
